refactor(*): put activities and 'triggers' in actions
...  which is what they're called in David Harel's original paper.
It makes the model more clear and removes the need to do all kinds of merge & squash tricks in the various renders to put the actions in the right order.

- This commit also tightens down the RE to recognize entry and exit actions - which was necessary because the previous one tended to eat stuff that became before a keyword.
- ... and it does away with the 'activities' top level key in favor of a actions of type 'activity'

all in all a BREAKING CHANGE w.r.t. the previous version of the AST (technically other commits in the same PR did the samem but it has to be flagged somewhere ...)