bugfix(test): fix a race condition
on the ci (where > 2 builds are running at the same time) we regularly get an exception that tmp_hello.json could not be found, confusingly returning a test failure on a completely different test. The only reason tmp_hello.json can't be there is because another process removed it. The most likely culprit is a test that is running in parallel (=> apparently the ci uses a shared file system?). This fix randomizes the filename so it's usually unique per test run. Results are only verifyable by monitoring ci behavior the coming N pushes.