perf!: lazy loads modules (#229)
## Description

- lazy loads modules on the interpreter on state-machine-cat.js.org. 
- Even though state-machine-cat-inpage is adapted as well, code
splitting isn't enabled there for now; don't turn too many cogs at the
same time. Maybe in a separate PR.
- does the same in the cli
- for this to work it's necessary the main smcat render function to
become async which is a 🚨 BREAKING CHANGE

> While we're at it we can modernize the main interface a bit (e.g.
replace the export default with regular functions and cash in on
deprecations (if any left (?))). Separate PR, though.

## Motivation and Context

Typically faster as there's less code to load. With some (external)
modules this will make quite a difference. E.g. not having to load 800kb
of wasm on the cli when the dot executable is already available will
speed up execution significantly. Same on the website.

## How Has This Been Tested?

- [x] green ci

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
- The contribution will be subject to [The MIT license](../LICENSE), and
I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](./CONTRIBUTING.md).